requirements.txt
PILLOW==7.1.1
numpy==1.18.2